+++ commands/update.ts
@@ -1,18 +1,24 @@
-import { getLatestVersion, parse } from "./deps.ts";
-import { RegistryCtor, RegistryUrl } from "./registry/utils.ts";
-import { DenoLand } from "./registry/denoland.ts";
-import { JsDelivr } from "./registry/jsdelivr.ts";
-import { Npm } from "./registry/npm.ts";
-import { GithubRaw } from "./registry/github.ts";
-import { GitlabRaw } from "./registry/gitlab.ts";
-import { Unpkg } from "./registry/unpkg.ts";
-import { Skypack } from "./registry/skypack.ts";
-import { EsmSh } from "./registry/esm.ts";
-import { NestLand } from "./registry/nestland.ts";
-import { Jspm } from "./registry/jspm.ts";
-import { Denopkg } from "./registry/denopkg.ts";
-import { PaxDeno } from "./registry/paxdeno.ts";
-import { Jsr } from "./registry/jsr.ts";
+import {
+  colors,
+  expandGlob,
+  getLatestVersion,
+  parse,
+  Spinner,
+} from "../deps.ts";
+import { RegistryCtor, RegistryUrl } from "../registry/utils.ts";
+import { DenoLand } from "../registry/denoland.ts";
+import { JsDelivr } from "../registry/jsdelivr.ts";
+import { Npm } from "../registry/npm.ts";
+import { GithubRaw } from "../registry/github.ts";
+import { GitlabRaw } from "../registry/gitlab.ts";
+import { Unpkg } from "../registry/unpkg.ts";
+import { Skypack } from "../registry/skypack.ts";
+import { EsmSh } from "../registry/esm.ts";
+import { NestLand } from "../registry/nestland.ts";
+import { Jspm } from "../registry/jspm.ts";
+import { Denopkg } from "../registry/denopkg.ts";
+import { PaxDeno } from "../registry/paxdeno.ts";
+import { Jsr } from "../registry/jsr.ts";
 
 const registries: RegistryCtor[] = [
   DenoLand,
@@ -33,6 +39,61 @@ const registries: RegistryCtor[] = [
 export interface UpdateOptions {
   // don't permanently edit files
   dryRun?: boolean;
+}
+
+export default async function run(files: string[], options: UpdateOptions) {
+  const spinner = new Spinner({ message: "Scanning files..." });
+  spinner.start();
+
+  const depFiles: string[] = [];
+
+  for (const arg of files.map((x) => x.toString())) {
+    for await (const file of expandGlob(arg)) {
+      depFiles.push(file.path);
+    }
+  }
+
+  if (depFiles.length === 0) {
+    spinner.stop();
+    console.error("No files found.");
+    return;
+  }
+
+  if (depFiles.length === 1) {
+    spinner.message = `Updating dependencies of ${depFiles[0]}...`;
+  } else {
+    spinner.message = `Updating dependencies of ${depFiles.length} files...`;
+  }
+
+  const results: UpdateResult[] = [];
+
+  await Promise.all(depFiles.map(async (filename) => {
+    results.push(...await update(filename, options));
+  }));
+
+  spinner.stop();
+
+  const alreadyLatest = results.filter((x) => x.newVersion === undefined);
+
+  if (alreadyLatest.length > 0) {
+    console.log(colors.bold("\nAlready latest version:"));
+    for (const a of alreadyLatest) {
+      console.log(colors.dim(a.initUrl), "==", a.initVersion);
+    }
+  }
+
+  const updated = results.filter((x) => x.newVersion !== undefined);
+
+  if (updated.length > 0) {
+    console.log(
+      colors.bold(
+        options.dryRun ? "\nAble to update:" : "\nSuccessfully updated:",
+      ),
+    );
+    for (const s of updated) {
+      console.log(colors.green(s.initUrl), s.initVersion, "->", s.newVersion);
+    }
+  }
 }
 
 export interface UpdateResult {