+++ deps.ts
@@ -1,11 +1,11 @@
-export { parseArgs } from "https://deno.land/std@0.222.1/cli/parse_args.ts";
-export { expandGlob } from "https://deno.land/std@0.222.1/fs/expand_glob.ts";
-export * as colors from "https://deno.land/std@0.222.1/fmt/colors.ts";
-export { Spinner } from "https://deno.land/std@0.222.1/cli/spinner.ts";
-import { parse as parseVersion } from "https://deno.land/std@0.222.1/semver/parse.ts";
-import { format } from "https://deno.land/std@0.222.1/semver/format.ts";
-import { compare } from "https://deno.land/std@0.222.1/semver/compare.ts";
-import { type SemVer } from "https://deno.land/std@0.222.1/semver/types.ts";
+export { parseArgs } from "https://deno.land/std@0.223.0/cli/parse_args.ts";
+export { expandGlob } from "https://deno.land/std@0.223.0/fs/expand_glob.ts";
+export * as colors from "https://deno.land/std@0.223.0/fmt/colors.ts";
+export { Spinner } from "https://deno.land/std@0.223.0/cli/spinner.ts";
+import { parse as parseVersion } from "https://deno.land/std@0.223.0/semver/parse.ts";
+import { format } from "https://deno.land/std@0.223.0/semver/format.ts";
+import { compare } from "https://deno.land/std@0.223.0/semver/compare.ts";
+import { type SemVer } from "https://deno.land/std@0.223.0/semver/types.ts";
 
 export interface Version extends SemVer {
   prefix?: string;